srcs/auth-service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
cryptography==41.0.7
argon2-cffi==23.1.0
python-decouple==3.8
gunicorn==21.2.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
